requirements.txt
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
python-dotenv==1.0.0
twilio==8.10.3
openai==1.6.1
schedule==1.2.0
datetime
json5==0.9.14
